Another way to refactor dependency requirement
