+++ server/domain/category.js
@@ -1,9 +1,9 @@
 'use strict';
 
 var _ = require('underscore');
-var config = require('./../../config/config.json');
-var acl = require('./../acl');
-var audit = require('./../audit');
+var config = require('config');
+var acl = require('../acl');
+var audit = require('../audit');
 var etcdBaseUrl = 'http://' + config.etcdHost + ':' + config.etcdPort + '/v2/keys/';
 
 var getCategory = function (id, name, description, columns, features) {
