+++ server/etcd.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var config = require('./../config/config.json');
+var config = require('config');
 var Etcd = require('node-etcd');
 
 module.exports.client = new Etcd(config.etcdHost, config.etcdPort);
