+++ server/routes/authorisationRoutes.js
@@ -4,7 +4,7 @@ var etcd = require('../etcd');
 var _ = require('underscore');
 var acl = require('../acl');
 var validator = require('validator');
-var config = require('./../../config/config.json');
+var config = require('config');
 
 module.exports = {
     getUsers: function (req, res) {
